research-app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-google-genai==1.0.7
langchain-core==0.2.10
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
yt-dlp==2024.4.9